Make new android video service the default
