+++ src/services/index.ts
@@ -1,6 +1,6 @@
 export { default as AndroidScreenshotService } from './android-screenshot.service'
+export { default as AndroidVideoLegacyService } from './android-video-legacy.service'
 export { default as AndroidVideoService } from './android-video.service'
-export { default as AndroidVideoAltService } from './android-video-alt.service'
 export { default as ConfigService } from './config.service'
 export { default as DeviceService } from './device.service'
 export { default as FfmpegService } from './ffmpeg.service'
